requirements.txt
numpy
pathlib
pillow
tensorflow
tensorflow-datasets
keras
flask
pylint
pylint-exit